requirements.txt
fastapi
uvicorn
pymupdf
openai
python-multipart
requests
python-dotenv